drm/i915: Sanitize LVDS state upon KMS takeover
We associate various auxiliary state like the panel fitter with the LVDS
encoder. So if the LVDS encoder is switched off by the BIOS, but the
panel fitter is left on, we never try to turn off the panel fitter and
leave it still attached to the pipe - which can cause blurry output
elsewhere.

Reported-by: andreas.sturmlechner@gmail.com
Bugzilla: https://bugs.freedesktop.org/show_bug.cgi?id=58867
Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>