drm/i915: track request boundaries
We used request->head only to track the flushing region
of request for post hang analysis.

Expand the use of request->head so that it tracks the
actual head of everything that is related the request
in question.

Request requeuing after hang will need this to know
request boundaries inside ringbuffer.

Signed-off-by: Mika Kuoppala <mika.kuoppala@intel.com>