drm/i915: initialize gem and ring seqnos separately
When requests are going to be replayed, we need to init
ring seqnos to the values that were completed before reset
happened and thus gem's notion of next seqno to what
was the last seqno in request queue + 1.

In preparation to this, split seqno initialization so that
we can init both parts separately.

Signed-off-by: Mika Kuoppala <mika.kuoppala@intel.com>