drm/i915: replay requests after reset
We add intel_ring_replay for rings to store the replay state.
Before reset we store ring tail, completed seqno and possibly
the hanged_seqno for post reset analysis.

After reset has been done, we go thru each ring to see
if guilty batchbuffer/context was found. If not, we
fall back to clearing the request queues and skip replay.

If guilty was found, request queues are analysed and
if request was guilty of hang or if request belongs to
banned context, we overwrite the ring commands that would
jump into a batch with MI_NOOP.

i915_gem_replay_requests() returns the largest seqno
found in request_queues or zero if replay preparation
was not done.

If replay preparation was done, we set per ring seqnos
to match the situation before the reset. Then we init
the gems notion of next seqno to match the pre reset situation,
and start the rings from the pre-reset positions.

Signed-off-by: Mika Kuoppala <mika.kuoppala@intel.com>