drm/i915: introduce intel_batchbuffer helpers
For initial render state generation we need a batch
which contains the commands and also the states for
various units in 3d pipeline.

We will construct the batch state so that the commands
are emitted to the start of the batch and states are
setup in the bottom half of the batch.

If allocation error for batch states occur during
state construction, wrong offsets will be returned
for the state, but the command emission side
will not push these offsets if error is set.

The reason for having this kind of error handling
is to utilize i-g-t state generators without minimal
modifications.

Signed-off-by: Mika Kuoppala <mika.kuoppala@intel.com>